% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmptbl_set.R
\name{comp_tbl}
\alias{comp_tbl}
\title{Get harmonies for data sets}
\usage{
comp_tbl(.data, lgran, ugran, ...)
}
\arguments{
\item{.data}{a tsibble object}

\item{lgran}{lower granularity}

\item{ugran}{gran granularity}

\item{...}{added arguments to be passed}
}
\value{
compatibility table providing if the two granularities are harmonies or clashes. FALSE indicates a clash. If harmony, then a tibble with desired granularities returned.
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(dplyr)
library(tsibbledata)
aus_elec \%>\% comp_tbl(lgran = "hour", ugran = "week")
}
