% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatible.R
\name{compatibility}
\alias{compatibility}
\title{Get compatibility tables for two granularities}
\usage{
compatibility(.data, ...)
}
\arguments{
\item{data}{a tsibble object}

\item{fn1}{the first granularity function to use}

\item{fn2}{the first granularity function to use}

\item{argfn1}{granularity to be fed in fn1}

\item{argfn2}{granularity to be fed in fn2}
}
\value{
compatibility table providing if the two granularities are harmonies or clashes. It also provides information on the range of the number of observations per combination and variation across number of combinations and other summery statistics.
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
\dontrun{
tsibbledata::aus_elec \%>\% dplyr::mutate(hour_day = ghour(Time, "day"), day_week = gday(Time, "week")) \%>\% compatibility("hour_day", "day_week")
}
}
