% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_norm_pairwise_max.R
\name{compute_norm_pairwise_max}
\alias{compute_norm_pairwise_max}
\title{Title Compute normalised maximum pairwise distances}
\usage{
compute_norm_pairwise_max(
  .data,
  gran_x = NULL,
  gran_facet = NULL,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  nperm = 100,
  lambda = 0.67
)
}
\arguments{
\item{.data}{data for which mmpd needs to be calculated}

\item{gran_x}{granularities mapped across x levels}

\item{gran_facet}{granularities mapped across facetss}

\item{response}{univarite response variable}

\item{quantile_prob}{probabilities}

\item{dist_ordered}{if categories are ordered}

\item{nperm}{number of permutations required for normalisation}

\item{lambda}{value of the tuning parameter for computing weighted pairwise distances}
}
\value{
the weighted pairwise distance normalised using permutation
}
\description{
Title Compute normalised maximum pairwise distances
}
\examples{
library(tidyverse)
library(gravitas)
library(parallel)
sm <- smart_meter10 \%>\%
  filter(customer_id \%in\% c("10017936"))
gran_x <- "week_month"
gran_facet <- "week_fortnight"
v <- compute_norm_pairwise_max(sm, gran_x, gran_facet,
  response = general_supply_kwh, nperm = 10
)
# month of the year not working in this setup
}
