% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pairwise_dist.R
\name{compute_pairwise_dist}
\alias{compute_pairwise_dist}
\title{Title computing all within and between facet distances between quantile categories given a data}
\usage{
compute_pairwise_dist(
  .data,
  gran_x = NULL,
  gran_facet = NA,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  lambda = 0.67
)
}
\arguments{
\item{.data}{data for which mmpd needs to be calculated}

\item{gran_x}{granularities mapped across x levels}

\item{gran_facet}{granularities mapped across facets}

\item{response}{univarite response variable}

\item{quantile_prob}{probabilities}

\item{dist_ordered}{if categories are ordered}

\item{lambda}{value of tuning parameter for computing weighted pairwise distances}
}
\value{
the raw weighted pairwise  within-facet and between-facet distances
}
\description{
Title computing all within and between facet distances between quantile categories given a data
}
\examples{
library(tidyverse)
library(gravitas)
library(parallel)
sm <- smart_meter10 \%>\%
  filter(customer_id \%in\% c("10017936"))
gran_x <- "month_year"
gran_facet <- "wknd_wday"
v <- compute_pairwise_dist(sm, gran_x, gran_facet,
  response = general_supply_kwh
)
# month of the year not working in this setup
}
