% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quantiles.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{compute quantiles of data across categories}
\usage{
compute_quantiles(sim_panel_data, quantile_prob = seq(0.01, 0.99, 0.01))
}
\arguments{
\item{sim_panel_data}{data with categories and response variable}

\item{quantile_prob}{quantiles of reponse variable needed}
}
\value{
data with quantiles of response variable corresponding to categories
}
\description{
compute quantiles of data across categories
}
\examples{
library(ggplot2)
library(dplyr)
library(distributional)
library(tidyr)
sim_panel_data <- sim_panel(
  nx = 3,
  nfacet = 2,
  ntimes = 100,
  sim_dist = distributional
  ::dist_normal(5, 10)
) \%>\%
  unnest(c(data))
sim_panel_data \%>\% ggplot() +
  geom_boxplot(aes(x = as.factor(id_x), y = sim_data)) +
  facet_wrap(~id_facet)
compute_quantiles(sim_panel_data) \%>\%
  unnest(sim_data_quantile) \%>\%
  ggplot() +
  geom_boxplot(aes(x = as.factor(id_x), y = unlist(sim_data_quantile))) +
  facet_wrap(~id_facet)

sim_varx_normal <- function(nx, nfacet, mean, sd, w) {
  rep(dist_normal((mean + seq(0, nx - 1, by = 1) * w), sd), nfacet)
}
data <- sim_panel(
  nx = 3,
  nfacet = 2,
  ntimes = 100,
  sim_dist = sim_varx_normal(nx = 3, nfacet = 2, mean = 0, sd = 1, w = 100)
)
sim_panel_data \%>\% ggplot() +
  geom_boxplot(aes(x = as.factor(id_x), y = sim_data)) +
  facet_wrap(~id_facet)
compute_quantiles(sim_panel_data) \%>\%
  unnest(sim_data_quantile) \%>\%
  ggplot() +
  geom_boxplot(aes(x = as.factor(id_x), y = unlist(sim_data_quantile))) +
  facet_wrap(~id_facet)
}
\author{
Sayani07
}
