% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gran.R
\name{create_gran}
\alias{create_gran}
\title{Build temporal granularities}
\usage{
create_gran(.data, gran1 = NULL, label = TRUE, abbr = TRUE, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{gran1}{Temporal granularity required.}

\item{label}{Logical. TRUE will display the month as an ordered factor of character string such as "January", "February". FALSE will display the month as an ordered factor such as 1 to 12, where 1 stands for January and 12 for December.}

\item{abbr}{logical. FALSE will display abbreviated labels}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tsibble with an additional column of granularity
}
\description{
Build temporal granularities
}
\examples{
library(dplyr)
library(tsibble)
tsibbledata::vic_elec \%>\% as_tsibble() \%>\% create_gran("hour_week") \%>\% tail()
}
