% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_harmony_data.R
\name{create_harmony_data}
\alias{create_harmony_data}
\title{Create data for one harmony pair}
\usage{
create_harmony_data(.data, harmony_tbl_row, response)
}
\arguments{
\item{.data}{a tsibble}

\item{harmony_tbl_row}{a row from the harmony table}

\item{response}{univariate response variable}
}
\value{
a tibble with the harmony pair and response variable
}
\description{
Create data for one harmony pair
}
\examples{
library(gravitas)
library(dplyr)
sm <- smart_meter10 \%>\%
  filter(customer_id \%in\% c("10017936"))
harmonies <- sm \%>\%
  harmony(
    ugran = "month",
    filter_in = "wknd_wday",
    filter_out = c("hhour", "fortnight")
  )
#harmonies <- harmonies \%>\% mutate(facet_variable = NA)
panel_data <- create_harmony_data(
  sm,
  harmonies[3, ], general_supply_kwh
)
}
