% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_harmony_tbl_data.R
\name{create_harmony_tbl_data}
\alias{create_harmony_tbl_data}
\title{Create data for all harmony pairs in a harmony table}
\usage{
create_harmony_tbl_data(.data, harmony_tbl = NULL, response = NULL)
}
\arguments{
\item{.data}{a tsibble}

\item{harmony_tbl}{harmony table}

\item{response}{univariate response variable}
}
\value{
a list with each element containing a tibble with the harmony pair and response variable
}
\description{
Create data for all harmony pairs in a harmony table
}
\examples{
library(gravitas)
library(dplyr)
sm <- smart_meter10 \%>\%
  dplyr::filter(customer_id \%in\% c("10017936"))
harmonies <- sm \%>\%
  harmony(
    ugran = "month",
    filter_in = "wknd_wday",
    filter_out = c("hhour", "fortnight")
  )
# harmonies <- harmonies \%>\% mutate(facet_variable = NA)
all_harmony <- create_harmony_tbl_data(sm,
  harmony_tbl = harmonies[1:7, ],
  response = general_supply_kwh
)
}
