% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cricket}
\alias{cricket}
\title{Cricket data IPL season 2008 to 2016}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 51360 rows and 10 columns.}
\usage{
data(cricket)
}
\description{
The Indian Premiere League played
by teams representing
different cities in India conisting of match_id, batting and bowling #' team, innings, over, runs per overs and rate rate per over
The data is obtained from url{https://www.kaggle.com/littleraj30
/indian-premier-league-2019-ball-by-ball/version/1}
}
\keyword{datasets}
