% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cricketdata}
\alias{cricketdata}
\title{Cricket data for IPL 2008}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 56640 rows and 24 columns.}
\usage{
data(cricketdata)
}
\description{
The Indian Premiere League played
by eight teams representing
different cities in India for 2008.
The data is obtained from url{https://www.kaggle.com/littleraj30/indian-premier-league-2019-ball-by-ball/version/1}
}
\keyword{datasets}
