% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_create_gran.R
\name{dynamic_create_gran}
\alias{dynamic_create_gran}
\title{Build dynamic temporal granularities}
\usage{
dynamic_create_gran(.data, hierarchy_tbl = NULL, gran = NULL,
  verify_col = FALSE, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{hierarchy_tbl}{A hierarchy table}

\item{gran}{the required granularity.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tsibble with an additional column of granularity
}
\description{
Build dynamic temporal granularities
}
\examples{
library(dplyr)
library(tsibble)
tsibbledata::vic_elec \%>\% as_tsibble() \%>\% create_gran("hour_week") \%>\% tail()
}
