% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gday.R
\name{gday}
\alias{gday}
\title{Get combination of day component of a date time
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.}
\usage{
gday(x, granularity = "week", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with day}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the day component of x as a number
}
\description{
Get combination of day component of a date time
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(ggplot2)
library(dplyr)
tsibbledata::aus_elec \%>\% mutate(day_week = gday(Time, "week")) \%>\% tail()
gday(lubridate::now(), "month")
}
