% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfortnight.R
\name{gfortnight}
\alias{gfortnight}
\title{Get combination of fortnight component of a date time}
\usage{
gfortnight(x, granularity = "month")
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with fortnight}
}
\value{
combination of the week component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
\dontrun{
tsibbledata::aus_elec \%>\% mutate(ft_year = gfortnight(Time, "year")) \%>\% tail()
gfortnight(lubridate::now(), "month")
}
}
