% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghhour.R
\name{ghhour}
\alias{ghhour}
\title{Get combination of half-hour component of a date time}
\usage{
ghhour(x, granularity = "hour", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with half-hour}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the hour component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(ggplot2)
library(dplyr)
tsibbledata::aus_elec \%>\% mutate(hh_day = ghhour(Time, "day")) \%>\% tail()
ghhour(lubridate::now(), "week")
}
