% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghour.R
\name{ghour}
\alias{ghour}
\title{Get combination of hour component of a date time}
\usage{
ghour(x, granularity = "day", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with hour}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the hour component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(ggplot2)
library(dplyr)
aus_elec <- tsibbledata::aus_elec \%>\%
  mutate(
    hour_day = ghour(Time, "day"),
    month_year = gmonth(Time, "year")
  )
aus_elec \%>\%
  filter(
    hour_day \%in\% c(1, 10, 15),
    month_year \%in\% c(4, 5, 6)
  ) \%>\%
  ggplot(aes(x = as.factor(hour_day), y = Demand)) + facet_wrap(~month_year) +
  geom_boxplot()
aus_quant <- aus_elec \%>\%
  filter(hour_day \%in\% c(1, 10, 15), month_year \%in\% c(4, 5, 6)) \%>\%
  group_by(hour_day, month_year) \%>\%
  do({
    x <- .$Demand
    purrr::map_dfr(
      .x = c(0.1, 0.5, 0.9),
      .f = ~ tibble(
        Quantile = .x,
        Value = quantile(x, probs = .x, na.rm = TRUE)
      )
    )
  })
aus_quant \%>\% ggplot(aes(x = hour_day, y = Value, col = as.factor(Quantile))) + geom_line() + facet_wrap(~month_year)
}
