% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_threshold.R
\name{global_threshold}
\alias{global_threshold}
\title{Selecting harmonies with significant difference in distributions for two cyclic granularities}
\usage{
global_threshold(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL,
  step1_data = NULL,
  ...
)
}
\arguments{
\item{.data}{a tsibble.}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}

\item{step1_data}{If the data across all harmony pairs available in list}
}
\description{
Selecting harmonies with significant difference in distributions for two cyclic granularities
}
\examples{
\dontrun{
library(tsibbledata)
library(ggplot2)
library(tsibble)
library(lvplot)
library(dplyr)
library(gravitas)
library(purrr)
library(magrittr)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10017936"))
.data = sm
harmonies <- sm \%>\%
harmony(ugran = "month",
       filter_in = "wknd_wday",
       filter_out = c("hhour", "fortnight"))
gran1 = "wknd_wday"
gran2 = "hour_day"
response  = "general_supply_kwh"
global_harmony <-  sm \%>\%
global_threshold(harmony_tbl = harmonies,
response = "general_supply_kwh")
}
}
