% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gminute.R
\name{gminute}
\alias{gminute}
\title{Get combination of minutes component of a date time}
\usage{
gminute(x, granularity = "hour", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with minutes}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the hour component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
tsibbledata::nyc_bikes \%>\% mutate(m_hour = gminute(Time, "hour")) \%>\% tail()
gminute(lubridate::now(), "day")
}
