% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmonth.R
\name{gmonth}
\alias{gmonth}
\title{Get combination of month component of a date time}
\usage{
gmonth(x, granularity = "year", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with month}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the month component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
\dontrun{
tsibbledata::aus_elec \%>\% mutate(month_year = gmonth(Time, "year")) \%>\% tail()
gmonth(lubridate::now(), "quarter")
}
}
