% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqhour.R
\name{gqhour}
\alias{gqhour}
\title{Get combination of quarter of an hour component of a date time}
\usage{
gqhour(x, granularity = "day")
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with  quarter of an hour}
}
\value{
combination of the quarter hour component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
\dontrun{
tsibbledata::aus_elec \%>\% mutate(qhour_day = gqhour(Time, "hour")) \%>\% tail()
gqtrhour(lubridate::now(), "day")
}
}
