% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gquarter.R
\name{gquarter}
\alias{gquarter}
\title{Get combination of quarter component of a date time}
\usage{
gquarter(x, granularity = "year", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with quarter}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the quarter component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
\dontrun{
tsibbledata::aus_elec \%>\% mutate(quarter_year = gquarter(Time, "year")) \%>\% tail()
gquarter(lubridate::now(), "semester")
}
}
