% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gran_advice.R
\name{gran_advice}
\alias{gran_advice}
\title{Recommendations on plot choices and summary of interaction, number of observations and intra or inter facet homogeneity}
\usage{
gran_advice(.data, gran1, gran2, hierarchy_tbl, response = NULL, ...)
}
\arguments{
\item{.data}{a tsibble}

\item{gran1, gran2}{granularities}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units and their relationships}

\item{response}{response variable to be plotted}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
a list of summary check points before visualizing distribution across bivariate granularities
}
\description{
Recommendations on plot choices and summary of interaction, number of observations and intra or inter facet homogeneity
}
\examples{
library(gravitas)
library(tsibbledata)
library(tsibble)
vic_elec \%>\% gran_advice(.data, gran1="hour_week",
gran2 = "day_month")
}
