% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granplot.R
\name{granplot}
\alias{granplot}
\title{Get combination of granularities of a date time
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.}
\usage{
granplot(.data, ugran = NULL, lgran = NULL, response = NULL,
  plot_type = NULL, ...)
}
\arguments{
\item{.data}{a tsibble}

\item{ugran}{the upper level granularity to be paired}

\item{lgran}{the lower level granularity to be paired}

\item{response}{response variable to be plotted}

\item{plot_type}{type of distribution plot.}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of granularities of x as a number
}
\description{
Get combination of granularities of a date time
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(dplyr)
library(tsibble)
tsibbledata::aus_elec \%>\%as_tsibble() \%>\% tail() \%>\%  create_gran("hour", "week")
}
