% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsecond.R
\name{gsecond}
\alias{gsecond}
\title{Get combination of seconds component of a date time}
\usage{
gsecond(x, granularity = "hour", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with seconds}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the seconds component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
\dontrun{
tsibbledata::nyc_bikes \%>\% mutate(sec_hour = gsecond(start_time, "hour")) \%>\% tail()
gsecond(lubridate::now(), "day")
}
}
