% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsemester.R
\name{gsemester}
\alias{gsemester}
\title{Get combination of semester component of a date time}
\usage{
gsemester(x, granularity = "year", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with semester}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the semester component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
tsibbledata::aus_elec \%>\% mutate(semester_year = gsemester(Time, "year")) \%>\% tail()
}
