% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gweek.R
\name{gweek}
\alias{gweek}
\title{Get combination of week component of a date time}
\usage{
gweek(x, granularity = "month")
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with week}
}
\value{
combination of the week component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(ggplot2)
library(dplyr)
tsibbledata::aus_elec \%>\% mutate(week_year = gweek(Time, "year")) \%>\% tail()
gweek(lubridate::now(), "month")
}
