% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_distance.R
\name{hist_distance}
\alias{hist_distance}
\title{Plotting probability distributions across granularities}
\usage{
hist_distance(
  .data = NULL,
  gran1 = NULL,
  gran2 = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL,
  dist_distribution = "normal"
)
}
\arguments{
\item{.data}{a tsibble.}

\item{gran1}{the granularity which is to be placed across facets. Can be column names if required granularity already exists in the tsibble. For example, a column with public holidays which needs to be treated as granularity, can be included here.}

\item{gran2}{the granularity to be placed across x-axis. Can be column names if required granularity already exists in the tsibble.}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}

\item{dist_distribution}{Underlying distribution of distances. Look at hist_distance()}
}
\value{
A tibble of harmonies and their levels ranked ion descending order of average maximum pairwise distance of the harmony pairs.
}
\description{
Plot probability distribution of univariate series across bivariate temporal granularities.
}
\examples{
library(tsibbledata)
library(ggplot2)
library(tsibble)
library(lvplot)
library(dplyr)
library(gravitas)
library(purrr)
library(magrittr)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c(10017936))
.data = sm
response  = "general_supply_kwh"
.data \%>\% hist_distance(gran1 = "hour_day", gran2 = "wknd_wday", response = "general_supply_kwh")
}
