% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPD1.R
\name{MMPD1}
\alias{MMPD1}
\title{Normalise max pairwise distance for one variable}
\usage{
MMPD1(
  harmony_datai,
  quantile_prob = seq(0.01, 0.99, by = 0.01),
  dist_distribution = "normal",
  dist_ordered = TRUE,
  base = exp(1),
  ...
)
}
\arguments{
\item{harmony_datai}{a tibble with many rows whose column consists of lists of numeric response variable}

\item{quantile_prob}{numeric vector of probabilities with values in [0,1].}

\item{dist_distribution}{Underlying distribution of distances.}

\item{dist_ordered}{if levels of the time granularity is ordered.}

\item{base}{base of logarithm while normalizing}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
MMPD for one harmony pair
}
\description{
Normalise max pairwise distance for one variable
}
