% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R, R/recursive_nest.R
\name{nest}
\alias{nest}
\title{Get combination of hour component of a date time}
\usage{
nest(gran1, gran2, x, ...)

nest(gran1, gran2, x, ...)
}
\arguments{
\item{gran1}{the lower level granularity to be paired}

\item{gran2}{the upper level granularity to be paired}

\item{x}{a date-time object}

\item{...}{other arguments to be passed for appropriate labels}

\item{granularity1}{granularity level 1}

\item{granularity2}{any higher order granularity to be paired with granularity 1}

\item{x}{a date-time object}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the hour component of x as a number

combination of granularities of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(ggplot2)
library(dplyr)
aus_elec <- tsibbledata::aus_elec \%>\%
  mutate(
    hour_day = nest(Time, "quarter", "year")
  )
library(dplyr)
tsibbledata::nyc_bikes \%>\% tail() \%>\% mutate(hhour_week = nest("hhour", "week", start_time))
}
