% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_jsd_maxharmony.R
\name{norm_jsd_maxharmony}
\alias{norm_jsd_maxharmony}
\title{Normalise max pairwise distance for one variable}
\usage{
norm_jsd_maxharmony(
  harmony_datai,
  quantile_prob = seq(0.01, 0.99, by = 0.01),
  dist_distribution = "normal",
  dist_ordered = TRUE
)
}
\arguments{
\item{harmony_datai}{a tibble with one row whose column consists of lists of numeric response variable}

\item{quantile_prob}{numeric vector of probabilities with values in [0,1].}

\item{dist_distribution}{Underlying distribution of distances.}

\item{dist_ordered}{if levels of the time granularity is ordered.}
}
\value{
normalised max JS distance
}
\description{
Normalise max pairwise distance for one variable
}
