% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_max_JSdist.R
\name{normalise_max_JSdist}
\alias{normalise_max_JSdist}
\title{Normalise max pairwise distance for one variable}
\usage{
normalise_max_JSdist(
  harmony_datai,
  quantile_prob = seq(0.01, 0.99, by = 0.01),
  dist_distribution = "normal",
  dist_ordered = TRUE
)
}
\arguments{
\item{harmony_datai}{a tibble with one row whose column consists of lists of numeric response variable}

\item{quantile_prob}{numeric vector of probabilities with values in [0,1].}

\item{dist_distribution}{Underlying distribution of distances.}

\item{dist_ordered}{if levels of the time granularity is ordered.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
normalised max JS distance
}
\description{
Normalise max pairwise distance for one variable
}
