% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gran_advice.R
\name{print.gran_advice}
\alias{print.gran_advice}
\title{Advice summaries for granularities}
\usage{
\method{print}{gran_advice}(x, ...)
}
\arguments{
\item{x}{An object of class gran_advice}

\item{...}{other arguments to be passed for appropriate labels.}
}
\value{
Print check points before visualizing distribution across bivariate granularities
}
\description{
Recommendations on plot choices, interaction, number of observations and intra or inter facet homogeneity. Important summaries before drawing distribution plots.
}
\examples{
library(dplyr)
library(ggplot2)

smart_meter10 \%>\%
filter(customer_id == "10017936") \%>\%
  gran_advice(gran1 = "wknd_wday",
            gran2 = "hour_day")

}
