% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_harmonies.R
\name{rank_harmonies}
\alias{rank_harmonies}
\title{rank harmonies with significant patterns}
\usage{
rank_harmonies(
  .data,
  harmony_tbl = NULL,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  lambda = 0.67,
  nperm = 20,
  use_perm = TRUE
)
}
\arguments{
\item{.data}{a tsibble or data with already computed categories}

\item{harmony_tbl}{A tibble containing one or more hamronies with facet_variable, x_variable, facet_levels and x_levels}

\item{response}{the response variable}

\item{quantile_prob}{numeric vector of probabilities with value #'in [0,1]  whose sample quantiles are wanted. Default is set to #' "decile" plot}

\item{dist_ordered}{if categories are ordered}

\item{lambda}{value of tuning parameter for computing weighted pairwise distances}

\item{nperm}{number of permutations for normalization}

\item{use_perm}{should permutation approach for normalization be used}
}
\description{
rank harmonies with significant patterns
}
\examples{
library(gravitas)
library(parallel)
library(dplyr)
library(tidyr)
sm <- smart_meter10 \%>\%
  filter(customer_id \%in\% c("10017936"))
harmonies <- sm \%>\%
  harmony(
    ugran = "month",
    filter_in = "wknd_wday",
    filter_out = c("hhour", "fortnight")
  )
all_harmony <- rank_harmonies(sm,
  harmony_tbl = harmonies,
  response = general_supply_kwh
)
}
