% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_harmony.R
\name{rank_harmony}
\alias{rank_harmony}
\title{Plotting probability distributions across granularities}
\usage{
rank_harmony(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL
)
}
\arguments{
\item{.data}{a tsibble.}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}
}
\value{
A tibble of harmonies and their levels ranked ion descending order of average maximum pairwise distance of the harmony pairs.
}
\description{
Plot probability distribution of univariate series across bivariate temporal granularities.
}
\examples{
library(tsibbledata)
library(ggplot2)
library(tsibble)
library(lvplot)
library(dplyr)
library(gravitas)
library(purrr)
library(magrittr)
harmonies <- smart_meter10 \%>\%
 harmony(ugran = "month",
         filter_out = c("hhour", "fortnight"))
.data = smart_meter10
response  = "general_supply_kwh"
harmony_tbl =  harmonies
.data \%>\% rank_harmony(harmony_tbl = harmonies, response = "general_supply_kwh")
harmony_tbl <- PBS \%>\% harmony(ugran = "year")
rank_harmony(PBS, harmony_tbl = harmony_tbl, response = "Cost")
}
