% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_harmony_v0.R
\name{rank_harmony_v0}
\alias{rank_harmony_v0}
\title{Plotting probability distributions across granularities}
\usage{
rank_harmony_v0(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  dist_distribution = "normal",
  hierarchy_tbl = NULL,
  dist_ordered = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{.data}{a tsibble.}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{dist_distribution}{Underlying distribution of distances. Look at hist_distance()}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}

\item{dist_ordered}{if levels of the time granularity is ordered.}

\item{alpha}{significance level}
}
\value{
A tibble of harmonies and their levels ranked in descending order of average maximum pairwise distance of the harmony pairs.
}
\description{
Plot probability distribution of univariate series across bivariate temporal granularities.
}
