% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_MMPD.R
\name{rank_MMPD}
\alias{rank_MMPD}
\title{Normalise max pairwise distance for one variable}
\usage{
rank_MMPD(
  .data,
  harmony_tbl = NULL,
  response = NULL,
  hierarchy_tbl = NULL,
  create_gran_data = TRUE,
  quantile_prob = seq(0.01, 0.99, by = 0.01),
  dist_distribution = "normal",
  dist_ordered = TRUE,
  ...
)
}
\arguments{
\item{.data}{a tsibble, if cyclic granularity needs to be constructed or a list consisting of tibbles for each pair of cyclic granularity in the harmony table}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}

\item{create_gran_data}{if data corresponding to a pair of cyclic granularity needs to be created}

\item{quantile_prob}{numeric vector of probabilities with values in [0,1].}

\item{dist_distribution}{Underlying distribution of distances.}

\item{dist_ordered}{if levels of the time granularity is ordered.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
MMPD for one harmony pair
}
\description{
Normalise max pairwise distance for one variable
}
