% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Runs a shiny app demonstrating functionalities of gravitas}
\usage{
run_app()
}
\value{
opens a local shiny app
}
\description{
Runs a local shiny app that demonstrates how distributions of univariate time series can be explored across bivariate time granularities
}
\examples{
\dontrun{
run_app()
}
}
\author{
Sayani Gupta
}
