% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_gran.R
\name{search_gran}
\alias{search_gran}
\title{Get set of temporal granularities for a given tsibble}
\usage{
search_gran(.data, ugran = "year", lgran = NULL, filter_in = NULL,
  filter_out = NULL, ...)
}
\arguments{
\item{.data}{a tsibble object}

\item{ugran}{upper temporal unit}

\item{lgran}{lowest temporal unit}

\item{filter_in}{choices of temporal units to be kept}

\item{filter_out}{choices of temporal units to be discarded}

\item{...}{added arguments to be passed}
}
\value{
set of granularities
}
\description{
Get set of temporal granularities for a given tsibble
}
\examples{
library(dplyr)
library(tsibbledata)
tsibbledata::vic_elec \%>\% search_gran(ugran = "month", filter_out = c("hhour","fortnight"))
tsibbledata::gafa_stock \%>\% search_gran(ugran = "month",lgran = "week")
}
