% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_distharmony1.R
\name{sim_distharmony1}
\alias{sim_distharmony1}
\title{Selecting harmonies with significant difference in distributions for two cyclic granularities}
\usage{
sim_distharmony1(
  .data,
  ntimes = 500,
  sim_dist = distributional::dist_normal(mu = 2, sigma = 5),
  ...
)
}
\arguments{
\item{.data}{a tsibble.}

\item{ntimes}{the size of the sample}

\item{sim_dist}{the distribution to be generated using package distributional}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Selecting harmonies with significant difference in distributions for two cyclic granularities
}
\examples{
library(ggplot2)
library(gravitas)
library(purrr)
library(distributional)
library(magrittr)
library(tidyr)
harmonies <- tibble::tibble(
  facet_variable =
    c("A", "B"),
  x_variable = c("B", "A"), facet_levels = c(2, 3), x_levels = c(3, 2)
)
.data <- harmonies[1, ]
data <- sim_distharmony1(.data,
  sim_dist = c(
    rep(dist_normal(mu = 2, sigma = 5), 3),
    rep(dist_normal(mu = 5, sigma = 10), 3)
  )
)
}
