% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sm_cust50}
\alias{sm_cust50}
\title{Smart meter data for 50 households}
\format{An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1278870 rows and 8 columns.}
\usage{
data(sm_cust50)
}
\description{
Customer Trial data conducted as part of
Smart Grid Smart City (SGSC) project
(2010-2014) based in Newcastle,
New South Wales
and areas in Sydney.
It contains half hourly interval meter readings (KWh)
of electricity consumption
and generation of households.
The data is obtained from url{https://data.gov.au/dataset/ds-dga-4e21dea3-9b87-4610-94c7-15a8a77907ef/details?q=smart-meter}
}
\keyword{datasets}
