% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smart_meter10}
\alias{smart_meter10}
\title{Smart meter data for ten households}
\format{A tsibble with 259, 235 rows and 3 columns:
* **customer_id**: household ID
* **reading_datetime**: Date time for which data is recorded (index)
* **general_supply_kwh**:electricity supplied to this household}
\usage{
data(smart_meter10)
}
\description{
Customer Trial data conducted as part of
Smart Grid Smart City (SGSC) project
(2010-2014) based in Newcastle,
New South Wales
and areas in Sydney.
It contains half hourly interval meter readings (KWh)
of electricity consumption and generation of households.
}
\references{
[Department of the Environment and Energy, Australia](https://data.gov.au/dataset/ds-dga-4e21dea3-9b87-4610-94c7-15a8a77907ef/details?q=smart-meter)
}
\keyword{datasets}
