% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_meter_13_cust2-data.R,
%   R/smart_meter_13_cust3-data.R
\docType{data}
\name{smart_meter_13_cust2}
\alias{smart_meter_13_cust2}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17464 rows and 4 columns.}
\source{
\href{https://data.gov.au/dataset/ds-dga-4e21dea3-9b87-4610-94c7-15a8a77907ef/details?q=smart-meter}{QTL Archive}

\href{https://data.gov.au/dataset/ds-dga-4e21dea3-9b87-4610-94c7-15a8a77907ef/details?q=smart-meter}{QTL Archive}
}
\usage{
data(smart_meter_13_cust2)

data(smart_meter_13_cust2)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.

Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\references{
(\href{https://data.gov.au/dataset/ds-dga-4e21dea3-9b87-4610-94c7-15a8a77907ef/details?q=smart-meter}{PubMed})

(\href{https://data.gov.au/dataset/ds-dga-4e21dea3-9b87-4610-94c7-15a8a77907ef/details?q=smart-meter}{PubMed})
}
\keyword{datasets}
