% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_harmonies.R
\name{threshold_harmonies}
\alias{threshold_harmonies}
\title{Selecting threshold with rcompanion package}
\usage{
threshold_harmonies(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL
)
}
\arguments{
\item{.data}{a tsibble.}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}
}
\description{
Selecting threshold with rcompanion package
}
