% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_create_gran.R
\name{validate_gran}
\alias{validate_gran}
\title{Validate created granularities with existing columns}
\usage{
validate_gran(.data, hierarchy_tbl = NULL, gran = NULL,
  validate_col = NULL, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{hierarchy_tbl}{A hierarchy table}

\item{gran}{the granularity to be created for validation.}

\item{validate_col}{A column in the data which acts as validator}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tsibble with an additional column of granularity
}
\description{
Validate created granularities with existing columns
}
\examples{
library(dplyr)
library(tsibble)
tsibbledata::vic_elec \%>\% as_tsibble() \%>\% create_gran("hour_week") \%>\% tail()
}
