% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_threshold.R
\name{wpd_threshold}
\alias{wpd_threshold}
\title{chooses threshold for a harmony table}
\usage{
wpd_threshold(
  .data,
  harmony_tbl = NULL,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  lambda = 0.67,
  nperm = 20,
  use_perm = TRUE,
  probs = c(0.9, 0.95, 0.99),
  nsamp = 100,
  create_harmony_data = TRUE
)
}
\arguments{
\item{.data}{a tsibble or data with already computed categories}

\item{harmony_tbl}{A tibble containing one or more hamronies with facet_variable, x_variable, facet_levels and x_levels}

\item{response}{the response variable}

\item{quantile_prob}{numeric vector of probabilities with value #'in [0,1]  whose sample quantiles are wanted. Default is set to #' "decile" plot}

\item{dist_ordered}{if categories are ordered}

\item{lambda}{value of tuning parameter for computing weighted}

\item{nperm}{number of permutations for normalization}

\item{use_perm}{should permutation approach for normalization be used}

\item{probs}{threshold probability}

\item{nsamp}{number of samples considered to compute threshold}

\item{create_harmony_data}{a logical value indicating if data corresponding to harmonies to be created or not}
}
\description{
chooses threshold for a harmony table
}
\examples{
library(gravitas)
library(parallel)
library(dplyr)
library(tidyr)
sm <- smart_meter10 \%>\%
  filter(customer_id \%in\% c("10017936"))
harmonies <- sm \%>\%
  harmony(
    ugran = "month",
    filter_in = "wknd_wday",
    filter_out = c("hhour", "fortnight")
  )
#all_harmony <- wpd_threshold(sm,
  #harmony_tbl = harmonies,
  #response = general_supply_kwh, nsamp = 3
#)
}
