% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatible.R
\name{compatibility}
\alias{compatibility}
\title{Get compatibility tables for two granularities}
\usage{
compatibility(.data, gran1, gran2, response = NULL, ...)
}
\arguments{
\item{gran1}{the first granularity function to use}

\item{gran2}{the first granularity function to use}

\item{response}{variable for which summary is desired per combination}

\item{...}{added arguments to be passed}

\item{data}{a tsibble object}
}
\value{
compatibility table providing if the two granularities are harmonies or clashes. It also provides information on the range of the number of observations per combination and variation across number of combinations and other summery statistics.
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(dplyr)
library(tsibbledata)
aus_elec \%>\%  compatibility("hour_day", "day_week")
}
