% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_nest.R
\name{create_gran}
\alias{create_gran}
\title{Get combination of granularities of a date time
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.}
\usage{
create_gran(.data, gran1 = NULL, gran2 = NULL, ...)
}
\arguments{
\item{.data}{a tsibble}

\item{gran1}{the lower level granularity to be paired}

\item{gran2}{the upper level granularity to be paired}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of granularities of x as a number
}
\description{
Get combination of granularities of a date time
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(dplyr)
library(tsibble)
tsibbledata::aus_elec \%>\%as_tsibble() \%>\% tail() \%>\%  create_gran("hour", "week")
}
