% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_create_gran.R
\name{dynamic_create_gran}
\alias{dynamic_create_gran}
\title{Build dynamic temporal granularities}
\usage{
dynamic_create_gran(.data, gran1 = NULL, hierarchy_tbl = NULL,
  label = TRUE, abbr = TRUE, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{gran1}{the granularity to be created}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units and their relationships}

\item{label}{Logical. TRUE will display the month as an ordered factor of character string such as "January", "February". FALSE will display the month as an ordered factor such as 1 to 12, where 1 stands for January and 12 for December.}

\item{abbr}{logical. FALSE will display abbreviated labels}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tsibble with an additional column of granularity
}
\description{
Build dynamic temporal granularities
}
\examples{
library(dplyr)
library(tsibble)
cricket_tsibble <- cricketdata \%>\%
mutate(data_index = row_number()) \%>\%
as_tsibble(index = data_index)
hierarchy_model <- tibble::tibble(units = c("index", "ball", "over", "inning", "match"),
convert_fct  = c(1, 6, 20, 2, 1))
cricket_tsibble \%>\% dynamic_create_gran("ball_over", hierarchy_model)
}
