% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghour.R
\name{ghour}
\alias{ghour}
\title{Get combination of hour component of a date time}
\usage{
ghour(x, granularity = "day", ...)
}
\arguments{
\item{x}{a date-time object}

\item{granularity}{the granularity to be paired up with hour}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
combination of the hour component of x as a number
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
tsibbledata::aus_elec \%>\% mutate(hour_day = ghour(Time, "day")) \%>\% tail()
ghour(lubridate::now(), "week")
}
