% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_harmony.R
\name{gran_obs}
\alias{gran_obs}
\title{Cross tabulation of granularities
useful for validating if number of observations are sufficient to compute probability distributions}
\usage{
gran_obs(.data, gran1, gran2, hierarchy_tbl = NULL, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{gran1}{One of the temporal granularities to check for harmonies.}

\item{gran2}{The second temporal granularity in the pair.}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units and their relationships}

\item{...}{Added arguments to be passed}
}
\value{
TRUE if two granularties are harmonies.
}
\description{
Cross tabulation of granularities
useful for validating if number of observations are sufficient to compute probability distributions
}
\examples{
library(dplyr)
library(tsibbledata)
library(ggplot2)
vic_elec \%>\% gran_obs("hour_day", "day_week")
}
