% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granplot.R
\name{granplot}
\alias{granplot}
\title{Plot probability distribution of univariate series across bivariate temporal granularities}
\usage{
granplot(.data, gran1 = NULL, gran2 = NULL, response = NULL,
  plot_type = NULL, facet_h = 31, ...)
}
\arguments{
\item{.data}{a tsibble}

\item{gran1}{the lower level granularity to be paired}

\item{gran2}{the upper level granularity to be paired}

\item{response}{response variable to be plotted}

\item{plot_type}{type of distribution plot.}

\item{facet_h}{levels of facet variable for which facetting is allowed while plotting bivariate temporal granularities}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
a ggplot object
}
\description{
Plot probability distribution of univariate series across bivariate temporal granularities
}
\examples{
library(dplyr)
library(tsibbledata)
vic_elec \%>\% granplot("hour_day", "day_week", "Demand", plot_type = "boxplot")
}
