% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granplot.R
\name{granplot}
\alias{granplot}
\title{Plot probability distribution of univariate series across bivariate temporal granularities}
\usage{
granplot(.data, gran1 = NULL, gran2 = NULL, hierarchy_tbl = NULL,
  response = NULL, plot_type = NULL, quantile_prob = c(0.1, 0.25,
  0.5, 0.75, 0.9), facet_h = NULL, overlay = TRUE, alpha = 0.8, ...)
}
\arguments{
\item{.data}{a tsibble}

\item{gran1}{the lower level granularity to be paired}

\item{gran2}{the upper level granularity to be paired}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units and their relationships}

\item{response}{response variable to be plotted}

\item{plot_type}{type of distribution plot}

\item{quantile_prob}{numeric vector of probabilities with value in [0,1]  whose sample quantiles are wanted. Default is set to "Decile" plot}

\item{facet_h}{levels of facet variable for which facetting is allowed while plotting bivariate temporal granularities}

\item{overlay}{If FALSE, only lines are drawn for quantiles. If TRUE, area is drawn instead of lines}

\item{alpha}{level of transperancy for the quantile area}

\item{...}{other arguments to be passed for appropriate labels}
}
\value{
a ggplot object
library(gravitas)
library(tsibbledata)
library(dplyr)
library(tsibble)
library(ggplot2)
vic_elec %>% granplot("hour_day", "day_week", response = "Demand", plot_type = "boxplot")

cricket_tsibble <- cricketdata %>%
  mutate(data_index = row_number()) %>%
  as_tsibble(index = data_index)

hierarchy_model <- tibble::tibble(
  units = c("index", "ball", "over", "inning", "match"),
  convert_fct = c(1, 6, 20, 2, 1)
)

cricket_tsibble %>%
  granplot("inning_match", "ball_inning",
    hierarchy_model,
    response = "total_runs",
    plot_type = "quantile",
    quantile_prob = c(0.1, 0.25, 0.5, 0.75, 0.9)
  )


granplot(
  .data = tsibbledata::vic_elec,
  gran1 = "day_week",
  gran2 = "hour_day",
  response = "Demand",
  plot_type = "quantile",
  quantile_prob = c(0.01, 0.1, 0.25, 0.5, 0.75, 0.9, 0.99),
  overlay = FALSE
)
}
\description{
Plot probability distribution of univariate series across bivariate temporal granularities
}
