% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmptbl_set.R
\name{harmony}
\alias{harmony}
\title{Get set of harmonies for a given tsibble}
\usage{
harmony(.data, ugran = NULL, lgran = NULL, ...)
}
\arguments{
\item{.data}{a tsibble object}

\item{ugran}{upper granularity}

\item{lgran}{lower granularity}

\item{...}{added arguments to be passed}
}
\value{
compatibility table providing if the two granularities are harmonies or clashes. FALSE indicates a clash. If harmony, then a tibble with desired granularities returned.
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(dplyr)
library(tsibbledata)
tsibbledata::gafa_stock \%>\% harmony(lgran = "hour", ugran = "week")
tsibbledata::vic_elec \%>\% harmony(ugran = "day")
}
