% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmony}
\alias{harmony}
\title{Get possible set of harmonies for a given tsibble}
\usage{
harmony(.data, ugran = "year", lgran = NULL, filter_in = NULL,
  filter_out = NULL, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{ugran}{Upper temporal unit. Typically, it is set as the most coarse temporal unit required in the analysis. Default is "year".}

\item{lgran}{Lowest temporal unit. For "regular" tsibble, lgran is the interval of the tsibble. It needs to be specified for "irregular" time intervals.}

\item{filter_in}{Choices of temporal units to be kept.}

\item{filter_out}{Choices of temporal units to be discarded.}

\item{...}{added arguments to be passed}
}
\value{
compatibility table providing if the two granularities are harmonies or clashes. FALSE indicates a clash. If harmony, then a tibble with desired granularities returned.
}
\description{
Get possible set of harmonies for a given tsibble
}
\examples{
library(dplyr)
library(tsibbledata)
tsibbledata::vic_elec \%>\% harmony(ugran = "day")
}
