% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.harmony.R
\name{is.harmony}
\alias{is.harmony}
\title{Check if two temporal granularities are harmonies}
\usage{
is.harmony(.data, gran1, gran2, hierarchy_tbl = NULL, response = NULL,
  facet_h = NULL, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{gran1}{One of the temporal granularities to check for harmonies.}

\item{gran2}{The second temporal granularity in the pair.}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units and their relationships}

\item{response}{Variable for which summary is desired per combination}

\item{facet_h}{levels of facet variable for which facetting is allowed while plotting bivariate temporal granularities.}

\item{...}{Added arguments to be passed}
}
\value{
TRUE if two granularties are harmonies.
}
\description{
Check if two temporal granularities are harmonies
}
\examples{
library(dplyr)
library(tsibbledata)
library(ggplot2)
vic_elec \%>\% is.harmony("hour_day", "day_week")
}
