% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatible.R
\name{is.harmony}
\alias{is.harmony}
\title{Get compatibility tables for two granularities}
\usage{
is.harmony(.data, gran1, gran2, response = NULL, ...)
}
\arguments{
\item{.data}{a tsibble object}

\item{gran1}{the first granularity function to use}

\item{gran2}{the first granularity function to use}

\item{response}{variable for which summary is desired per combination}

\item{...}{added arguments to be passed}
}
\value{
compatibility table providing if the two granularities are harmonies or clashes. FALSE indicates a clash. If harmony, then a tibble with desired granularities returned.
}
\description{
Date-time must be a  POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo,
zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.
}
\examples{
library(dplyr)
library(tsibbledata)
library(ggplot2)
aus_elec \%>\% is.harmony("hour_day", "day_week")
}
