% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_harmony.R
\name{rank_harmony}
\alias{rank_harmony}
\title{Plotting probability distributions across granularities}
\usage{
rank_harmony(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  dist_distribution = "normal",
  hierarchy_tbl = NULL,
  dist_ordered = TRUE,
  alpha = 0.05,
  create_gran_data = TRUE
)
}
\arguments{
\item{.data}{a tsibble, if cyclic granularity needs to be constructed or a list consisting of tibbles for each pair of cyclic granularity in the harmony table}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{dist_distribution}{Underlying distribution of distances. Look at hist_distance()}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}

\item{dist_ordered}{if levels of the time granularity is ordered.}

\item{alpha}{significance level}

\item{create_gran_data}{if data corresponding to a pair of cyclic granularity needs to be created}
}
\value{
A tibble of harmonies and their levels ranked in descending order of average maximum pairwise distance of the harmony pairs.
}
\description{
Plot probability distribution of univariate series across bivariate temporal granularities.
}
\examples{
library(tsibbledata)
library(ggplot2)
library(tsibble)
library(lvplot)
library(dplyr)
library(gravitas)
library(purrr)
library(magrittr)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10017936"))
harmonies <- sm \%>\%
harmony(ugran = "month",
       filter_in = "wknd_wday",
       filter_out = c("hhour", "fortnight"))
.data = sm
response  = "general_supply_kwh"
harmony_tbl =  harmonies
smart_harmony <- .data \%>\% rank_harmony(harmony_tbl = harmonies,
response = "general_supply_kwh", dist_ordered = TRUE)
harmony_tbl <- PBS \%>\% harmony(ugran = "year")
rank_harmony(PBS, harmony_tbl = harmony_tbl, response = "Cost")
}
