% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{groups}
\alias{as_tibble}
\alias{tibble}
\alias{!!}
\alias{!!!}
\alias{quo_name}
\alias{:=}
\alias{xlab}
\alias{ylab}
\alias{scale_fill_brewer}
\alias{aes}
\alias{geom_lv}
\alias{quantile}
\alias{median}
\alias{shinytheme}
\alias{useShinyalert}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{groups}}}

  \item{ggplot2}{\code{\link[ggplot2]{xlab}}, \code{\link[ggplot2]{ylab}}, \code{\link[ggplot2]{scale_fill_brewer}}, \code{\link[ggplot2]{aes}}}

  \item{lvplot}{\code{\link[lvplot]{geom_lv}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}, \code{\link[rlang]{quo_name}}, \code{\link[rlang]{:=}}}

  \item{shinyalert}{\code{\link[shinyalert]{useShinyalert}}}

  \item{shinythemes}{\code{\link[shinythemes]{shinytheme}}}

  \item{stats}{\code{\link[stats]{quantile}}, \code{\link[stats]{median}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}
}}

