% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_gran.R
\name{search_gran}
\alias{search_gran}
\title{Get set of possible temporal granularities for a given tsibble}
\usage{
search_gran(.data, ugran = "year", lgran = NULL, filter_in = NULL,
  filter_out = NULL, ...)
}
\arguments{
\item{.data}{A tsibble object.}

\item{ugran}{Upper temporal unit. Typically, it is set as the most coarse temporal unit required in the analysis. Default is "year".}

\item{lgran}{Lowest temporal unit. For "regular" tsibble, lgran is the interval of the tsibble. It needs to be specified for "irregular" time intervals.}

\item{filter_in}{Choices of temporal units to be kept.}

\item{filter_out}{Choices of temporal units to be discarded.}

\item{...}{Other arguments to be passed.}
}
\value{
Set of possible granularities.
}
\description{
Get set of possible temporal granularities for a given tsibble
}
