% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_distharmony1.R
\name{sim_distharmony1}
\alias{sim_distharmony1}
\title{Selecting harmonies with significant difference in distributions for two cyclic granularities}
\usage{
sim_distharmony1(
  .data,
  ntimes = 500,
  sim_dist = distributional::dist_normal(mu = 2, sigma = 5),
  ...
)
}
\arguments{
\item{.data}{a tsibble.}

\item{ntimes}{the size of the sample}

\item{sim_dist}{the distribution to be generated using package distributional}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Selecting harmonies with significant difference in distributions for two cyclic granularities
}
