% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_new.R, R/threshold_harmonies.R
\name{threshold_harmonies}
\alias{threshold_harmonies}
\title{Selecting harmonies with significant difference in distributions for two cyclic granularities}
\usage{
threshold_harmonies(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL
)

threshold_harmonies(
  .data = NULL,
  harmony_tbl = NULL,
  response = NULL,
  prob = seq(0.01, 0.99, 0.01),
  hierarchy_tbl = NULL
)
}
\arguments{
\item{.data}{a tsibble.}

\item{harmony_tbl}{A tibble of harmonies and their levels obtained from the function().}

\item{response}{response variable.}

\item{prob}{numeric vector of probabilities with values in [0,1].}

\item{hierarchy_tbl}{A hierarchy table specifying the hierarchy of units}
}
\description{
Selecting harmonies with significant difference in distributions for two cyclic granularities

Selecting threshold with rcompanion package
}
\examples{
library(tsibbledata)
library(ggplot2)
library(tsibble)
library(lvplot)
library(dplyr)
library(gravitas)
library(purrr)
library(magrittr)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c(10017936))
.data = sm
gran1 = "wknd_wday"
gran2 = "hour_day"
response  = "general_supply_kwh"
}
